+++ tests/testthat/test-ggcoef_model.R
@@ -225,6 +225,11 @@ test_that("ggcoef_multinom()", {
   )
 
   vdiffr::expect_doppelganger(
+    "ggcoef_multinom() table",
+    ggcoef_multinom(mod, type = "table", exponentiate = TRUE)
+  )
+
+  vdiffr::expect_doppelganger(
     "ggcoef_multinom() faceted custom y level label",
     ggcoef_multinom(
       mod,
@@ -352,6 +357,10 @@ test_that("ggcoef_table()", {
     ggcoef_table(mod_simple, table_header = c("A", "B", "C"))
   )
 
+  expect_error(
+    ggcoef_table(mod_simple, table_header = c("A", "B", "C", "D"))
+  )
+
   vdiffr::expect_doppelganger(
     "ggcoef_table() table_text_size",
     ggcoef_table(mod_simple, table_text_size = 5)
@@ -388,6 +397,16 @@ test_that("ggcoef_table()", {
   )
 
   vdiffr::expect_doppelganger(
+    "ggcoef_table() show_p_values only",
+    ggcoef_table(mod_simple, show_p_values = TRUE, signif_stars = FALSE)
+  )
+
+  vdiffr::expect_doppelganger(
+    "ggcoef_table() signif_stars only",
+    ggcoef_table(mod_simple, show_p_values = FALSE, signif_stars = TRUE)
+  )
+
+  vdiffr::expect_doppelganger(
     "ggcoef_table() customized statistics",
     ggcoef_table(
       mod_simple,
@@ -403,5 +422,119 @@ test_that("ggcoef_table()", {
       table_header = c("Term", "Coef.", "SE", "CI"),
       table_witdhs = c(2, 3)
     )
+  )
+})
+
+test_that("ggcoef_multicomponents()", {
+  skip_on_cran()
+  skip_if_not_installed("broom.helpers")
+  skip_if_not_installed("pscl")
+
+  library(pscl)
+  data("bioChemists", package = "pscl")
+  mod <- zeroinfl(art ~ fem * mar | fem + mar, data = bioChemists)
+
+  vdiffr::expect_doppelganger(
+    "ggcoef_multicomponents() dodged",
+    ggcoef_multicomponents(mod)
+  )
+
+  vdiffr::expect_doppelganger(
+    "ggcoef_multicomponents() faceted",
+    ggcoef_multicomponents(mod, type = "f")
+  )
+
+  vdiffr::expect_doppelganger(
+    "ggcoef_multicomponents() table",
+    ggcoef_multicomponents(mod, type = "t")
+  )
+
+  expect_s3_class(
+    ggcoef_multicomponents(mod, type = "t", return_data = TRUE),
+    "tbl"
+  )
+
+  vdiffr::expect_doppelganger(
+    "ggcoef_multicomponents() table component_label",
+    ggcoef_multicomponents(
+      mod,
+      type = "t",
+      component_label = c(conditional = "Count", zero_inflated = "Zero-inflated") # nolint
+    )
+  )
+
+  vdiffr::expect_doppelganger(
+    "ggcoef_multicomponents() faceted component_label",
+    ggcoef_multicomponents(
+      mod,
+      type = "f",
+      component_label = c(conditional = "Count", zero_inflated = "Zero-inflated")  # nolint
+    )
+  )
+
+  # message if unfound values for component_label
+  expect_message(
+    ggcoef_multicomponents(
+      mod,
+      tidy_fun = broom.helpers::tidy_zeroinfl,
+      type = "t",
+      component_label = c(c = "Count", zi = "Zero-inflated")
+    )
+  )
+
+  # error if unnamed values for component_label
+  expect_error(
+    ggcoef_multicomponents(
+      mod,
+      tidy_fun = broom.helpers::tidy_zeroinfl,
+      type = "t",
+      component_label = c("Count", zi = "Zero-inflated")
+    )
+  )
+
+  mod2 <- zeroinfl(art ~ fem + mar | 1, data = bioChemists)
+  vdiffr::expect_doppelganger(
+    "ggcoef_multicomponents() mod2 table",
+    ggcoef_multicomponents(mod2, type = "t")
+  )
+
+  skip_if_not_installed("betareg")
+  skip_if_not_installed("parameters")
+
+  library(betareg)
+  data("GasolineYield", package = "betareg")
+  m1 <- betareg(yield ~ batch + temp, data = GasolineYield)
+  m2 <- betareg(yield ~ batch + temp | temp + pressure, data = GasolineYield)
+
+  vdiffr::expect_doppelganger(
+    "ggcoef_multicomponents() betareg m1 table",
+    ggcoef_multicomponents(
+      m1,
+      type = "t",
+      tidy_fun = broom.helpers::tidy_parameters,
+      tidy_args = list(component = "all")
+    )
+  )
+
+  vdiffr::expect_doppelganger(
+    "ggcoef_multicomponents() betareg m2 table",
+    ggcoef_multicomponents(
+      m2,
+      type = "t",
+      tidy_fun = broom.helpers::tidy_parameters,
+      tidy_args = list(component = "all")
+    )
+  )
+
+  modlm <- lm(Sepal.Length ~ Sepal.Width + Species, data = iris)
+
+  vdiffr::expect_doppelganger(
+    "ggcoef_multicomponents() linear model table",
+    ggcoef_multicomponents(modlm, type = "t")
+  )
+
+  vdiffr::expect_doppelganger(
+    "ggcoef_multicomponents() linear model faceted",
+    ggcoef_multicomponents(modlm, type = "f")
   )
 })